refactor: extract ploy-risk crate, unify risk management
Move slippage protection, validation chains, and circuit breaker into
a dedicated ploy-risk workspace crate. Add Round type to ploy-core.
Keep RiskManager adapter in main app (depends on platform::RiskGate).
Original modules re-export from ploy-risk for backward compatibility.

21 tests pass in ploy-risk, full suite (573 tests) green.
