+++ crates/ploy-risk/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "ploy-risk"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+ploy-core = { path = "../ploy-core" }
+rust_decimal = { version = "1", features = ["serde-with-str"] }
+rust_decimal_macros = "1"
+serde = { version = "1", features = ["derive"] }
+tracing = "0.1"
+chrono = { version = "0.4", features = ["serde"] }
+thiserror = "1"
+tokio = { version = "1", features = ["sync"] }
+
+[dev-dependencies]
+tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
